src/backend/expertmeet/requirements.txt
django
python-dotenv

djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

djangorestframework-simplejwt

django-cors-headers

Pillow